Implemented rf with different approach, removed components and used only future objects
